Fixed synchronization mechanism of MessageQueue: the construction of Continuation messages was not synchronized and therefore vulnerable to race conditions.
Also removed class MessageQueuedNotifier, since this would easily enable nested synchronization calls
(as it did when creating this fix): one on the message queue and one on the script executor.
This led to deadlocks.
The purpose of MessageQueuedNotifier was to kick the executor in case it was sleeping
and a new message had entered the queue.
However this was not necessary since such insertions would be done
using doCodeThatInsertsMsgs_synchronized.

This commit makes the synchronization a bit more basic and easier to spot.